\name{EPAbvs-package}
\alias{EPAbvs-package}
\alias{EPAbvs}
\docType{package}
\title{
\packageTitle{EPAbvs}
}
\description{
\packageDescription{EPAbvs}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{EPAbvs}
\packageIndices{EPAbvs}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{EPAbvs}

Maintainer: \packageMaintainer{EPAbvs}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
